Finished the main work on meson refactoring.
Lots of meson changes to make the app dynamically use the dev or release
app ids and prefixes.

Most notable:
I've moved gtk/ into data dir because that makes it easier to install
with Meson.

Gresources and schema and metainfo are configured
with the prefix and app id in meson to allow for separating dev builds
from release builds.

I've also moved tests outside of the src/ folder. I will work on
implementing some simple unit tests later. I first want to test that
everything in the app works as expected, including the data and
xdg paths.

I've moved most Meson configuration outside of src and into the top-level
dir to make it easier to parse everything out.
