+++ data/dev.qwery.AddWater.gschema.xml.in
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <schemalist gettext-domain="addwater">
 	<!-- Add Water -->
-	<schema id="@APP_ID@" path="@PREFIX@">
+	<!-- The slash at the end of the prefix is important! -->
+	<schema id="@APP_ID@" path="@PREFIX@/">
 		<key name="window-height" type="i">
 			<default>700</default>
 		</key>
@@ -85,16 +86,14 @@
 		</key>
 	</schema>
 
-
 	<!-- Thunderbird (finish when ready to support)-->
-	<schema id="@PREFIX@.Thunderbird" path="@PREFIX@/Thunderbird/">
+	<schema id="@APP_ID@.Thunderbird" path="@PREFIX@/Thunderbird/">
 		<key name="data-path" type="s">
 			<default>""</default>
 		</key>
 		<key name="autofind-paths" type="b">
 			<default>true</default>
 		</key>
-
 	</schema>
 </schemalist>
