+++ meson_options.txt
@@ -10,7 +10,7 @@ option(
 option(
 	'use-mock-api',
 	type: 'boolean',
-	value: true
+	value: false
 )
 
 option(
