--- src/addwater.gresource.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<gresources>
-	<gresource prefix="/dev/qwery/AddWater">
-		<file preprocess="xml-stripblanks">gtk/window.ui</file>
-		<file preprocess="xml-stripblanks">gtk/help-overlay.ui</file>
-		<file preprocess="xml-stripblanks">gtk/preferences.ui</file>
-		<file preprocess="xml-stripblanks">gtk/addwater-page.ui</file>
-	</gresource>
-
-	<gresource prefix="dev/qwery/AddWater/icons/scalable/actions">
-		<file preprocess="xml-stripblanks" alias="checkmark-symbolic.svg">../data/icons/hicolor/scalable/actions/checkmark-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="cross-large-symbolic.svg">../data/icons/hicolor/scalable/actions/cross-large-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="cross-small-symbolic.svg">../data/icons/hicolor/scalable/actions/cross-small-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="firefox-symbolic.svg">../data/icons/hicolor/scalable/actions/firefox-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="globe-symbolic.svg">../data/icons/hicolor/scalable/actions/globe-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="info-outline-symbolic.svg">../data/icons/hicolor/scalable/actions/info-outline-symbolic.svg</file>
-		<file preprocess="xml-stripblanks" alias="mail-symbolic.svg">../data/icons/hicolor/scalable/actions/mail-symbolic.svg</file>
-	</gresource>
-</gresources>
