+++ tests/test_install_manager.py
