+++ tests/test_online_manager.py
